Create SIT Clubs Application
The communication among college clubs is currently conducted through emails, phone calls, or word of mouth. However, these methods are inconvenient within an institution as large as SIT. To address this issue, we are in the process of developing a user-friendly software interface where all club-related updates, messages, news, and collaborative activities will be centralized. This platform will replace the conventional modes of communication, such as emails and physical posters, which are currently distributed through platforms like WhatsApp groups and others.